Alterando o tipo de 'price' do modelo de 'product'

+++ apps/product/models/product.py
@@ -12,7 +12,7 @@ class Category(models.Model):
 class Product(models.Model):
 
     name = models.CharField(max_length=70)
-    price = models.DecimalField(max_digits=7, decimal_places=2)
+    price = models.FloatField()
     criation_date = models.DateField(auto_now=True)
     category = models.ForeignKey(Category, models.DO_NOTHING)
     description = models.CharField(max_length=200)
